move: place-modal의 폴더 이동 (#10)
